Add correct home corner for blinky
