+++ levels/level_1.py
@@ -6,6 +6,8 @@ from settings import FPS
 from model.board_definition import BoardDefinition
 from model.level_config import LevelConfig
 
+# 32 x 30
+
 board = [
 [6, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5],
 [3, 6, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5, 6, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 5, 3],
